Use entities instead of units/items
Unit::Base and Item::Base are now subclasses of RubySurvivor::Entity.
This decision was made based on the fact that many parts of the
codebase don't need to know the difference betwen an item and a unit.
For example, in order for items to exist in the world (on a Floor)
like units, they must either be treated the same, or more branching
or methods must be added. There are already an uncomfortable number
of kind_of? calls in the codebase, and this will hopefully prevent
more from entering.